pyproject.toml
[tool.poetry]
name = "tomcamp_api"
version = "0.1.0"
description = ""
authors = ["Tom Camp <tomcamp@protonmail.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
gunicorn = "^20.1.0"
djangorestframework = "!=3.14.0"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.13.0"
whitenoise = "^6.2.0"
django-guardian = "^2.4.0"
django-filter = "^22.1"
drf-yasg = "^1.21.3"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pylint = "^2.15.3"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Pylint config
[tool.pylint.main]
recursive = true
load-plugins = ["pylint_django"]
ignore = ["migrations/*", "manage.py"]

[tool.pylint.format]
max-line-length = 120


[tool.pylint."messages control"]
disable = ["missing-module-docstring", "missing-class-docstring", "missing-function-docstring", "django-not-configured", "unspecified-encoding", "inconsistent-return-statements", "duplicate-code", "import-outside-toplevel"]
